Add files via upload
The aim of this project is to do a Data analysis on India's Air quality and predict the value of the air quality index based on the given features of concentration of sulpur dioxide, nitrogen dioxide, respirable suspended particulate matter, suspended particulate matter and classify the air quality as good, moderate, poor, unhealthy, healthy. The data is combined across the years and is a largely clean version of the Historical Daily Ambient Air Quality Data released by the Ministry of Environment and Forests and Central Pollution Control Board of India under the National Data Sharing and Accessibility Policy (NDSAP) year 1990-2015.
The classification accuracy has obtained the maximum for Random Forest classification.